requirements.txt
# ✅ 2. `requirements.txt` dosyası oluştur

requests
Pillow
fpdf
python-whois
dnspython
colorama
